// apps/shared/types/units.types.ts
export interface TroopTemplateData {
  slotIndex: number; // 0-4
  name: string;
  description?: string; // História/descrição opcional da tropa
  avatar?: string; // ID do sprite (ex: "1")
  passiveId: string;
  resourceType: keyof PlayerResources;
  combat: number;
  speed: number;
  focus: number;
  resistance: number;
  will: number;
  vitality: number;
}
// ...
export interface KingdomTemplateDefinition {
  id: string;
  name: string;
  description: string;
  alignment: Alignment;
  race: Race;
  regentCode: string;
  troopTemplates: TroopTemplateData[];
}